05-clothingController: createClothing

+++ develop/src/main/java/cz/cvut/fel/nss/vestana/repo/LoanRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LoanRepo extends CrudRepository<Loan, Integer> {
+public interface LoanRepo extends CrudRepository<Loan, Long> {
 }
